Add support for aarch64
Add support for aarch64 to the build system and nolibc. Note that libgcc.a is
required on aarch64 and will be automatically added to ocaml-freestanding.pc.

Support is restricted to building on Linux with gcc for now.

As part of this change I have removed the out of date and unused arm
configuration templates.