opam, Travis CI: Switch to Solo5 master for development
Now that Solo5/opam-solo5 is back online for development on master,
switch ocaml-freestanding to it by adding EXTRA_REMOTES to Travis CI and
updating OPAM metadata to depend on Solo5 >= 0.3.0.

As part of this, based on the earlier consensus on mirage-dev regarding
supported compiler versions and to speed up Travis CI cycles, remove
compilers older than 4.04.2 from the build matrix and update
ocaml-version in opam.