nolibc: Define __ocaml_freestanding__, add overrides.h
This change allows for portable C code to determine if it is being built
in an ocaml-freestanding environment by checking if the preprocessor
macro __ocaml_freestanding__ is defined.

In order to be able to do this reliably and in the presence of existing
"#if...#elif...#endif" chains we need to undefine existing preprocessor
macros identifying the host toolchain, including but not limited to,
__FreeBSD__, __OpenBSD__ and __linux__. For a full list, see
nolibc/include/_freestanding/overrides.h.

_freestanding/overrides.h is now included automatically for each C
source file built by adding "-include _freestanding/overrides.h" to
both downstream pkg-config/dune CFLAGS and the Makefile-internal
FREESTANDING_CFLAGS.