nolibc: Provide a sys/param.h with BYTE_ORDER
Provide a <sys/param.h> in nolibc, and define BYTE_ORDER, LITTLE_ENDIAN
and BIG_ENDIAN in it (if not already defined via inclusion of endian.h
or related elsewhere).

This change is provided for the benefit of autoconf-based build systems
such as the OCaml runtime and libgmp (gmp-freestanding).

Autoconf-based build systems commonly use the AC_C_BIGENDIAN test,
which looks for a definition of BYTE_ORDER in various UNIXy places [1],
falling back to gradually more horrible heuristics if one is not found.

While these heuristics work "by accident" until now, this solution is
cleaner and ensures that a compile-only test is used by autoconf, which
will work in the future presence of a cross compiler.

[1] See lib/autoconf/c.m4 in the GNU autoconf source.