WIP Create a OCaml/Solo5-specific cross toolchain
Add a script to generate all the tools of a OCaml/Solo5 specific
toolchain, using the `ARCH-solo5-ocaml-` tool prefix, in order to:

- allow to have a clean way to point to the build directories in `-I`
  and `-L` options for nolibc and openlibm during the build, and to the
  installed directories after installation
- build upon the toolchain autoconfiguration in OCaml's configure, so
  that it is not necessary to override each tool separately

Add Makefile recipes to build the two variants of the toolchain, the one
for build and the one to be installed

Add a configure option to set up the tool prefix to use for the binutils
that are not part of the Solo5 toolchain

WIP: This creates wrappers for all the standard binutils, OCaml is using
autoconf that detect most of them. This commit goes for consistency
(covering them all) but maybe some are really not necessary.