Generate an OPAM .install file
Add two scripts to create an .install file:
- `gen_ocaml_install.sh` drives (bends...) OCaml's `make install` in order
  to generate two subfiles, namely `_build/ocaml.install.lib{,exec}`,
  containing the lines to install the compiler itself, into the `lib`
  and `libexec` sections
- `gen_dot_install.sh` generates the complete `.install` file, including
  the content of the subfiles for the compiler

The generated `.install` file diverges from the current installation
process in that it tries to follow the standard organisation of the OPAM
directory: the OCaml cross-compiler is installed in `%{_:lib}%` (ie
`%{prefix}%/lib/ocaml-solo5`) instead of `%{prefix}%/solo5-sysroot`.
Note that the paths fed to `configure.sh` must be consistent with that
for the installed compiler to work as expected.